backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
markdown==3.5.1
psycopg2-binary==2.9.9
email-validator==2.0.0 